+++ repository/src/main/java/org/grnet/status/repositories/DowntimeRepository.java
@@ -8,8 +8,6 @@ import org.grnet.status.entities.PageQuery;
 import org.grnet.status.entities.PageQueryImpl;
 import org.grnet.status.entities.*;
 import org.grnet.status.entities.*;
-
-import org.grnet.status.entities.*;
 import org.grnet.status.entities.*;
 import java.time.Instant;
 import java.util.HashMap;